+++ client/src/app/auth/store/models/auth.state.model.ts
@@ -1,4 +1,4 @@
-import {User} from '../../user/model/user';
+import {User} from '../../../user/models/user';
 export interface AuthStateInterface {
 	isSubmitting: boolean,
 	user: User,
