+++ client/src/app/context-menu/context-menu.component.ts
@@ -123,8 +123,4 @@ export class ContextMenuComponent implements OnInit, AfterViewInit {
 			
 		}
 	}
-	
-
-	
-
 }
